requirements.txt
# Airflow is already in base image
spotipy==2.23.0
duckdb==0.10.0
pandas==2.1.4
python-dotenv==1.0.0
dbt-core==1.7.0
dbt-duckdb==1.7.0
requests==2.31.0
psycopg2-binary==2.9.9

# Slack notifications
apache-airflow-providers-slack==8.5.0

# OpenLineage for data lineage tracking
openlineage-airflow==1.8.0
openlineage-python==1.8.0